+++ src/Ivory/Connection/AutoTxHandle.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+namespace Ivory\Connection;
+
+/**
+ * Handle of an open transaction, rolling back the transaction automatically upon destruction if still open.
+ *
+ * {@inheritDoc}
+ *
+ * @see ITransactionControl::startAutoTransaction()
+ */
+class AutoTxHandle extends TxHandle
+{
+    public function __destruct()
+    {
+        $this->rollbackIfOpen();
+        parent::__destruct();
+    }
+}